requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
pillow==8.2.0
pytesseract==0.3.8
gunicorn==20.1.0
